+++ packages/package-json-lint-config-terra/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Added
+  * Added new rule `require-no-hard-coded-dependency-versions`.
+
 ## 1.0.0 - (August 13, 2021)
 
 * Initial stable release