+++ packages/package-json-lint/CHANGELOG.md
@@ -5,6 +5,9 @@
 ## 1.1.0 - (August 31, 2021)
 
 * Added
+  * Added new rule `require-no-hard-coded-dependency-versions`.
+
+* Added
   * Added `projectType` to allow users to flag project as `application`, `module` or `devModule`.
 
 ## 1.0.0 - (August 13, 2021)